requirements.txt
gradio>=6.0.0
plotly>=5.18.0
kaleido>=0.2.1
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
PyPortfolioOpt>=1.5.5
yfinance>=0.2.30
psycopg2-binary>=2.9.9
anthropic>=0.34.0
weasyprint>=60.0
python-dotenv>=1.0.0
